Fixing some compiler warnings and single-threaded game issues. Note: The server currently crashes when trying to start a game.

+++ src/core/common/avatarmanager.cpp
@@ -525,7 +525,7 @@ AvatarManager::RemoveOldAvatarCacheEntries()
 			while (!timeMap.empty() && !m_cachedAvatars.empty())
 			{
 				TimeAvatarMap::iterator i = timeMap.begin();
-				if (curTime - i->first < MAX_AVATAR_CACHE_AGE)
+				if (curTime - i->first < (int)MAX_AVATAR_CACHE_AGE)
 					break;
 				AvatarMap::iterator pos = m_cachedAvatars.find(i->second);
 				if (pos != m_cachedAvatars.end())
